feat: 마커용 가게 전체 조회 api 추가
issue: #3

+++ src/main/java/com/kinopio/eatgo/store/dto/StoreResponseDto.java
@@ -19,8 +19,8 @@ public class StoreResponseDto {
 	private Integer storeId;
 	private String storeName;
 	private String address;
-	private Integer postionX;
-	private Integer postionY;
+	private Double postionX;
+	private Double postionY;
 	private Integer isOpen;
 	
 	private Integer userId;
